Partial rollback of changelist 150805495.
Specifically, I'm rolling back only the server/java8 changes.

*** Reason for rollback ***

Need to investigate Bazel heap regression.

*** Original change description ***

Do not copy in Immutable{List,Set,SortedSet} construction.

We're still avoiding the potential for a memory leak with the forceCopy part, but we want to avoid the garbage created by slicing the array down to exactly the right size.

We don't ever copy in ImmutableList, since we can only be too big by a factor of 3/2.  In Immutable{Set,SortedSet}, we copy if, after deduplication, the number of unique elements is < array.length / 2.  (And we rebuild the hash table in IS if it's more than twice as...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151480429