Prepare common.graph for Java-8-specific changes.
It so happens that this makes a bunch of files no longer executable. They shouldn't have been in the first place, so that's good. It's not clear that we'll fix up our tools to stop making the rest of the files executable anytime soon, though.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159462625