Ensure that we generate a non-empty Javadoc jar for guava-gwt.
Without it, we get an error during Sonatype deployment:
  Missing: no javadoc jar found in folder '/com/google/guava/guava-gwt/28.2-jre'

I hacked around this on the GitHub release branch for 28.2 by reenabling Javadoc:
https://github.com/google/guava/commit/a1b3c06876803a0b0e5d2f16708e1328da1bac09

But as you may recall from CL 276327335, we're soon going to have no classes to generate Javadoc for (after we remove GWT-RPC support). So even with Javadoc generation enabled, we'd end up with no jar.

To ensure that we get a jar, I've introduced a package-private dummy class (and then excluded it from the other steps in which source files are used).

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288314285