Make nullWriter().append(...) (both overloads) accept a null CharSequence.
The docs of Writer are misleading:

"""
An invocation of this method of the form out.append(csq) behaves in exactly the same way as the invocation
     out.write(csq.toString())
"""
https://docs.oracle.com/javase/7/docs/api/java/io/Writer.html#append%28java.lang.CharSequence%29

But that's not true, as the docs go on to say:

"""
csq - The character sequence to append. If csq is null, then the four characters "null" are appended to this writer.
"""

Accepting null in the 2-arg method is arguably even weirder, but at least the docs call it out more prominently:
https://docs.oracle.com/javase/7/docs/api/java/io/Writer.html#append%28java.lang.CharSequence,%20int,%20int%29

Credit to the Checker Framework for identifying this bug.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=298388993