Split afterRanInterruptibly into afterRanInterruptiblySuccess and afterRanInterruptiblyFailure.
This makes the code easier to understand for nullness checkers -- and probably easier to understand, full stop.

(Also, rename the GWT emulation's InterruptibleTask's type parameter from V to T. This makes it match the normal version. And that in turn matches other classes like AbstractTransformFuture, which differentiates between V (the value type of the Future) and T (the type returned by the task, which may be either a V or a ListenableFuture<V>). While InterruptibleTask itself is not a Future, its subclasses are nested inside Future subclasses, so V may still be confusing. Plus, "T" is pretty much the "default" name for type parameters :))

RELNOTES=n/a
PiperOrigin-RevId: 371382377