Document when `EndpointPair.adjacentNode` changed signature.
(Followup to CL 371380054.)

As usual, the scheme is:

```
@since version-binary-compatible-since (other-clarifications)
```

And in fact this change is binary-compatible, since the erasure of `N` is `Object`, the old parameter type. And while this change was potentially source-incompatible, we knew that and tested for problems internally, and we saw no breakages.

Compare to [the `@since` tag for `Queues.newPriorityBlockingQueue`](https://github.com/google/guava/blob/1799bc3e48db5ea61ba66191096727d17d1e811c/guava/src/com/google/common/collect/Queues.java#L191).

RELNOTES=n/a
PiperOrigin-RevId: 372188501