Remove `Service.State.isTerminal()`.
It was only referenced in one place. Since we guarantee that transitions between the states in the `Service.State` enum are always from an earlier to a later value, we know that the terminal states must be last. So it's enough to compare with the first terminal state, `TERMINATED`. The code in question already does something very similar with the `RUNNING` state a few lines earlier.

RELNOTES=n/a
PiperOrigin-RevId: 376892852