Introduce `ImmutableMap.Builder.buildOrThrow()`.
This new method is equivalent to the existing `ImmutableMap.Builder.build()`. `build()` now just calls `buildOrThrow()` and will soon be deprecated. The idea is to make it more obvious that duplicate keys will cause an exception.

Make similar changes to the builders of `ImmutableBiMap`, `ImmutableSortedMap`, and `ImmutableTable`.

RELNOTES=The method `ImmutableMap.Builder.build()` is now renamed to `buildOrThrow()`. The existing `build()` method will continue to exist but will be deprecated and the new name should be used in new code.
PiperOrigin-RevId: 396865988