Remove `@Beta` from the `FluentFuture` type, its factory methods, and `addCallback`.
This leaves `@Beta` only for the methods that remain `@Beta` in `Futures` itself, which I'm hoping to make some effort toward later this week.

This CL is an admission that we're never going to get around to reworking `FluentFuture` to be more of a "builder," as discussed in https://github.com/google/guava/issues/3419. Maybe there's still a place for something more builder-like, but there's no chance that we'll try to change `FluentFuture` in place. And probably we'll never do anything.

RELNOTES=`util.concurrent`: Removed `@Beta` from the `FluentFuture` type, its factory methods, and `addCallback`.
PiperOrigin-RevId: 416122526