Update docs that refer to the diamond operator to reflect that it's no longer new.
In particular, all Guava users have been able to Java 7+ language features for a long time.

Note that I've left some other "Note for Java 7 and later" docs in place: Those docs refer to Java 7+ *APIs*, which are not necessarily available to our Android users. (Ideally, we would update the docs to refer to Android, but I'm trying not to turn this into a big project :))

RELNOTES=n/a
PiperOrigin-RevId: 420059386