Copy some tests from the mainline to the backport.
Those tests aren't necessarily all as _interesting_ for the backport's implementation, but they can be made to work with little or no modification. So a bit of extra testing shouldn't hurt, and it's nice to minimize diffs between the two versions.

Complications:
- `testBuilderPutAllEntryListJdkBacked` makes sense only when there is a JDK-backed implementation, so I didn't copy that over. (Probably there were other tests like this, too: This is just the one that I copied over without noticing the dependency on the JDK-backed implementation.)
- `testBuildKeepingLast_collisions` kind of only does, too, though maybe testing of hash collisions in general is a reasonable idea (though I want to say we already have testing for that...). I had to modify it slightly. I could be convinced to revert this one entirely.

I just happened to notice these while working to convert nested test classes to top-level classes.

(The tests came from cl/184313314 and cl/416035210.)

PiperOrigin-RevId: 438650986