Automatically add `@CheckForNull` annotations with https://github.com/google/error-prone/blob/master/core/src/main/java/com/google/errorprone/bugpatterns/nullness/ReturnMissingNullable.java
And convert existing `@Nullable` annotations to `@CheckForNull` where possible. (See "Nullness annotations" in https://github.com/google/guava/releases/tag/v31.0)

These two classes are two that we haven't tried to annotate fully, but we might as well annotate where it's trivial.

Note, though, that at least a few of the existing `@Nullable` annotations are likely to be incorrect. I don't plan to investigate, since this cleanup is increasingly not "trivial" :)

PiperOrigin-RevId: 479709282