More reliably use transitional `@CheckForNull` and `@ParametricNullness` annotations.
I would assume we're still not 100% consistent in using them, and we might someday consider an Error Prone check to identify the remaining places that we ought to be using them. But in an ideal world, they'll be able to go away entirely before we need to bother.

(There are a few reasons to use these annotations, but the one that I was poking at today was trying to [run NullAway](https://github.com/google/guava/issues/2484#issuecomment-1203200500).)

RELNOTES=n/a
PiperOrigin-RevId: 494880081