Fix a way of creating mutable `ImmutableSet` instances.
The `Sets.toImmutableSet()` collector did not clear its state when building the set, but the `EnumSet` it passes along is treated as immutable without copying. Nefarious direct callers could, with just an unchecked cast and abuse of the `Collector` API, keep adding to a 2+ element `ImmutableSet` of enums built this way.

We [do not in general defend against malicious callers](https://github.com/google/guava#important-warnings) (and this change is not a _complete_ defense anyway). But the fix here has essentially no complexity or performance cost, so we might as well do it.

RELNOTES=n/a
PiperOrigin-RevId: 499483029