Enable [Java8ApiChecker](https://errorprone.info/bugpattern/Java8ApiChecker) for Maven builds.
We already have it on for our builds inside Google, but Maven checking provides an additional line of defense against https://github.com/google/guava/issues/3990.

Because this CL enables checking for our _tests_, too (which is hard not to do, as discussed in the existing `pom.xml` comment about NullArgumentForNonNullParameter), I had to edit a couple benchmarks to avoid using what are _in a limited sense_ "Java 9+ APIs." They wouldn't be practical problems, though.

RELNOTES=n/a
PiperOrigin-RevId: 511830560