Add `@Beta` back to the new `Invokable.getAnnotatedReturnType` method of guava-android.
There is some chance that we will have to delete the method if we hear of problems under Android. And since it's a brand-new API _and_ it doesn't actually work under an Android VM, we know that we're affecting no existing users and possibly none even in the future.

RELNOTES=n/a
PiperOrigin-RevId: 523212269