Stop recognizing `@ParametricNullness` annotations in `NullPointerTester`.
Those annotations are temporary, so we haven't wanted to rely on them. We were relying on them anyway in `NullPointerTester` because we couldn't use type-use annotations at all in the backport and, even after we could use them, we couldn't rely on the backport's copy of `NullPointerTester` to recognize them. But those concerns are resolved as of cl/524904788.

This requires a workaround in one test to avoid a JDK bug. At that point, as you'd expect, our tests pass even if we remove `@ParametricNullness` entirely.

PiperOrigin-RevId: 530341970